refactor: revert NavItem type move to NavItems
