+++ polaris.shopify.com/src/components/Nav/index.ts
@@ -1,3 +1,4 @@
 import Nav from "./Nav";
+export type { NavItem } from "./Nav";
 
 export default Nav;
