+++ polaris.shopify.com/src/components/NavItems/index.ts
@@ -1,4 +1,3 @@
 import NavItems from "./NavItems";
-export type { NavItem } from "./NavItems";
 
 export default NavItems;
