+++ polaris.shopify.com/src/components/TokensPage/TokensPage.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import Container from "../Container";
 import { TokenPropertiesWithName } from "../../types";
 import TokenList from "../TokenList";
-import type { NavItem } from "../NavItems";
+import type { NavItem } from "../Nav";
 import Link from "next/link";
 
 interface Props {
